[MouseWithoutBorders] - moving Common.Event.cs -> Core\Event.cs - #35155
